Add user assets retrieval functionality
The commit introduces 'UserAssets' to the user struct and an 'Assets' method for fetching user's assets. The 'UserAssets' fields are span across the user's metadata. The 'Assets' function in users.go fetches the list of assets linked to a particular user ID. A new constant 'userAssetsAPI' has been added to map the API endpoint.